Bind non-SDL Number/Operator keys (N0..PERIOD)
The RPG2003 Numbers/Operators ids (RGSS::Input::N0..PERIOD, 21..35)
were previously only backed by the SDL desktop backend. Now the
terminal/sixel backends type them from a real keyboard, the PSP backend
binds N0-N4 to its spare Square/L/R/Start/Select buttons (its pad
bitmask widened to uint64_t since ids 21-35 no longer fit a 32-bit
mask), and the Wio Terminal reserves the id slots for parity but leaves
them unbound (no free pin). Input Number and the numeric/operator Key
Input Processing groups now work off-SDL where hardware permits.

+++ mruby-rgss/src/wio.cxx
@@ -95,11 +95,11 @@ void wio_input_init(void) {
     pinMode(p.pin, INPUT_PULLUP);
 }
 
-uint32_t wio_input_scan(void) {
-  uint32_t mask = 0;
+uint64_t wio_input_scan(void) {
+  uint64_t mask = 0;
   for (const PinMap& p : kPins) {
     if (digitalRead(p.pin) == LOW)  // active low
-      mask |= (1u << p.key);
+      mask |= (1ull << p.key);
   }
   return mask;
 }
